wrong contributors
Wrong contributors due to a contributor moving the files. The first version of Astar pushed bij user deniz112 got deleted and the file moved and added by someone else. Fixing this as I want to use the code for future usage without unjust plagiarism issues as I wrote the full code. No offense to the one that moved it.