Should have only 1 contributor (only deniz112 worked on this file). Second contributor only moved the file from removed folder.
Re-added file to attempt to fix contributor(s) being wrong due to moving of files. Wrong contributors due to a contributor moving the files. The first version of Astar pushed by user deniz112 got deleted and the file moved and added by someone else. Mentioning this as I want to use the code for future usage without unjust plagiarism issues as I wrote the full code. No offense to the one that moved it. Can fix by rewriting history, but better to stay at the safe side and just keep it like this.